% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seoul-package.R
\docType{package}
\name{seoul-package}
\alias{seoul}
\alias{seoul-package}
\title{seoul: Example Package for Seoul Short Course}
\description{
This is a cool package written by Daniel Sabanes.
}
\author{
\strong{Maintainer}: Daniel Sabanes \email{daniel@rpact.com}

Authors:
\itemize{
  \item Daniel Sabanes \email{daniel@rpact.com}
}

}
\keyword{internal}
